Remove call_with_correct_args()
We have started having some bugs come up due to our reliance on the call_with_correct_args() function. This commit removes the call_with_correct_args() function, and introduces a couple of new arguments to the subscribe function which are used to control what arguments the provided callback receives.